changed structures of directories
